backend/pyproject.toml
[project]
name = "ai-textbook-backend"
version = "1.0.0"
description = "RAG chatbot backend for AI textbook"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Your Name", email = "your@email.com"}]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "qdrant-client>=2.7.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    "langchain>=0.1.0",
    "openai>=1.0.0",
    "python-multipart>=0.0.6",
    "better-auth>=1.0.0",
    "pyjwt>=2.8.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "pylint>=3.0.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pylint.messages_control]
disable = ["C0103", "R0903"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
